+++ src/resources/index.ts
@@ -1,3 +1,4 @@
 export * from './ProjectMetadata';
+export * from './DraftProjectMetadata';
 export * from './FileMetadata';
 export * from './GrdmFiles';